requirements/common.txt
alembic==0.9.3
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-Migrate==2.0.4
Flask-SQLAlchemy==2.2
python-dateutil==2.6.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.16.0
SQLAlchemy==1.1.11
facebook-sdk==3.1.0